Tratando Validações

+++ src/main/java/com/wamk/cadastrarproduto/controllers/ProductController.java
@@ -6,6 +6,7 @@ import java.util.UUID;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,6 +20,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.wamk.cadastrarproduto.dto.ProductDTO;
 import com.wamk.cadastrarproduto.entities.Product;
 import com.wamk.cadastrarproduto.services.ProductService;
+
+import jakarta.validation.Valid;
 
 @RestController
 @RequestMapping("/products")
@@ -36,11 +39,14 @@ public class ProductController {
 	@GetMapping("/{id}")
 	public ResponseEntity<Product> findById(@PathVariable UUID id){
 		Optional<Product> productO = productService.findById(id);
+		if(productO.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 		return ResponseEntity.ok(productO.get());
 	}
 
 	@PostMapping
-	public ResponseEntity<Product> post(@RequestBody ProductDTO productDTO) {
+	public ResponseEntity<Product> post(@RequestBody @Valid ProductDTO productDTO) {
 		var product = new Product();
 		BeanUtils.copyProperties(productDTO, product);
 		product = productService.save(product);
@@ -48,9 +54,12 @@ public class ProductController {
 	}
 	
 	@PutMapping("/{id}")
-	public ResponseEntity<Object> post(@RequestBody ProductDTO productDTO,
+	public ResponseEntity<Object> post(@RequestBody @Valid ProductDTO productDTO,
 			@PathVariable UUID id) {
 		Optional<Product> productO = productService.findById(id);
+		if(productO.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 		BeanUtils.copyProperties(productDTO, productO.get());
 		productService.update(productO.get());
 		return ResponseEntity.ok(productO.get());
@@ -59,6 +68,9 @@ public class ProductController {
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Object> post(@PathVariable UUID id) {
 		Optional<Product> productO = productService.findById(id);
+		if(productO.isEmpty()) {
+			return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+		}
 		productService.delete(productO.get());
 		return ResponseEntity.noContent().build();
 	}
